Move string utils file into dedicated package
