+++ src/main/kotlin/Main.kt
@@ -1,4 +1,4 @@
-import util.getAuthorizationUrl
+import util.authorization.getAuthorizationUrl
 
 fun main() {
     println(getAuthorizationUrl())
